added allgames in html and view added a game model kevin
