+++ NerdQuest/events/templates/allgames.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+
+    <head>
+        <meta charset="UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=div, initial-scale=1.0">
+        <title>All Games</title>
+        {% load static%}
+        <link rel="stylesheet" href="{% static 'css/showstyle' %}">
+        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/5.0.0-alpha1/css/bootstrap.min.css"
+            integrity="sha384-r4NyP46KrjDleawBgD5tp8Y7UzmLA05oM1iAEQ17CSuDqnUK2+k9luXQOfXJCJ4I" crossorigin="anonymous">
+        <script src="https://stackpath.bootstrapcdn.com/bootstrap/5.0.0-alpha1/js/bootstrap.min.js"
+            integrity="sha384-oesi62hOLfzrys4LxRF63OJCXdXDipiYWBnvTl9Y9/TRlw5xlKIEHpNyvvDShgf/"
+            crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
+            integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
+            crossorigin="anonymous"></script>
+    </head>
+
+    <body>
+        <div class="container">
+            <table class="table table-dark">
+                <tr>
+                    <th>Type OF Game</th>
+                    <th>Date</th>
+                    <th>Start Time</th>
+                    <th>End Time</th>
+                    <th>Location</th>
+                    <th>Notes</th>
+                </tr>
+                {% for game in all_games %}
+                <tr>
+
+                    <td>{{game.type}}</td>
+                    <td>{{game.date}}</td>
+                    <td>{{game.start}}</td>
+                    <td>{{game.end}}</td>
+                    <td>{{game.location}}</td>
+                    <td>{{game.notes}}</td>
+                    <td><a href="/shows/{{show.id}}">Show</a> <a href="/shows/{{show.id}}/edit">Edit</a> <a href="/shows/{{show.id}}/destroy">Delete</a></td>
+
+                </tr>
+                {% endfor %}
+            </table>
+
+            <a href="/shows/new">Add A New Show</a>
+        </div>
+
+    </body>
+
+</html>
